Jupyter Notebook
Step#1: Install gtts library. Execute pip install gtts on Anaconda Shell, as we will be using Jupyter Notebook.

Step#2: On Jupyter Notebook, write the below code. The mentioned code demonstrates how a given text “you are a mad person” is converted into English automated voice. Also if we have a text which is not in English, but in any other language (let say in Hindi), gtts can also convert it well !! :)